// src/app/models/editproperty.request.ts
export interface editproperty{

    "id": number;
    "sellRent": number;
    "user": number;
    "name": string;
    "propertyTypeId": number;
    "furnishingTypeId": number;
    "price": number;
    "bhk": number;
    "builtArea": number;
    "cityId": number;
    "readyToMove": boolean;
    "carpetArea": number;
    "address1": string;
    "address2": string;
    "floorNo": number;
    "totalFloors": number;
    "mainEnterance": string;
    "security": number;
    "maintenanceAmount": number;
    "ageProperty": number;
    "description": string;
    "postedOn": string;
    "imageUrl": string;
  
    
}